Routes created from pins, species, users, and regulations moved to own file
